Split document storage service code into files
